LPS-60164 move packages
